+++ src/features/donations/index.tsx
@@ -6,8 +6,6 @@ import PaymentDetails from './screens/PaymentDetails';
 import ThankYou from './screens/ThankYou';
 import TreeDonation from './screens/TreeDonation';
 import { useAuth0 } from '@auth0/auth0-react';
-import { Elements } from '@stripe/react-stripe-js';
-import getStripe from '../../utils/stripe/getStripe';
 
 interface Props {
   onClose: any;
@@ -57,7 +55,7 @@ function DonationsPopup({
 
   const [token, setToken] = React.useState('');
 
-  const [userProfile,setUserprofile] = React.useState(null)
+  const [userProfile, setUserprofile] = React.useState(null)
 
   //  to load payment data
   React.useEffect(() => {
@@ -225,74 +223,64 @@ function DonationsPopup({
     }
   }, [directGift]);
 
-  if(paymentSetup){
-    switch (donationStep) {
-      case 1:
-        return (
-          <Elements 
-            stripe={getStripe(paymentSetup?.gateways?.stripe?.authorization.stripePublishableKey, paymentSetup?.gateways?.stripe?.authorization.accountId)}>
-            <motion.div
-              animate={{
-                scale: [0.94, 1.05, 1],
-              }}
-              transition={{ duration: 0.8 }}
-            >
-              <TreeDonation {...TreeDonationProps} />
-            </motion.div>
-          </Elements>
-        );
-      case 2:
-        return (
-          <motion.div
-            animate={{
-              scale: [0.94, 1.04, 1],
-            }}
-            transition={{ duration: 0.8 }}
-          >
-            <ContactDetails {...ContactDetailsProps} />
-          </motion.div>
-        );
-      case 3:
-        return (
-          
-            <motion.div
-              animate={{
-                scale: [0.94, 1.05, 1],
-              }}
-              transition={{ duration: 0.8 }}
-            >
-              <PaymentDetails {...PaymentDetailsProps} />
-            </motion.div>
-        );
-      case 4:
-        return (
-          <motion.div
-            animate={{
-              scale: [0.94, 1.04, 1],
-              rotate: [-15, 5, 0],
-            }}
-            transition={{ duration: 0.8 }}
-          >
-            <ThankYou {...ThankYouProps} />
-          </motion.div>
-        );
-      default:
-        return (
-          <Elements 
-            stripe={getStripe(paymentSetup?.gateways?.stripe?.authorization.stripePublishableKey, paymentSetup?.gateways?.stripe?.authorization.accountId)}>
-            <motion.div
-              animate={{
-                scale: [0.94, 1.05, 1],
-              }}
-              transition={{ duration: 0.8 }}
-            >
-              <TreeDonation {...TreeDonationProps} />
-            </motion.div>
-          </Elements>
-        );
-    }
-  }else{
-    return <></>;
+  switch (donationStep) {
+    case 1:
+      return (
+        <motion.div
+          animate={{
+            scale: [0.94, 1.05, 1],
+          }}
+          transition={{ duration: 0.8 }}
+        >
+          <TreeDonation {...TreeDonationProps} />
+        </motion.div>
+      );
+    case 2:
+      return (
+        <motion.div
+          animate={{
+            scale: [0.94, 1.04, 1],
+          }}
+          transition={{ duration: 0.8 }}
+        >
+          <ContactDetails {...ContactDetailsProps} />
+        </motion.div>
+      );
+    case 3:
+      return (
+
+        <motion.div
+          animate={{
+            scale: [0.94, 1.05, 1],
+          }}
+          transition={{ duration: 0.8 }}
+        >
+          <PaymentDetails {...PaymentDetailsProps} />
+        </motion.div>
+      );
+    case 4:
+      return (
+        <motion.div
+          animate={{
+            scale: [0.94, 1.04, 1],
+            rotate: [-15, 5, 0],
+          }}
+          transition={{ duration: 0.8 }}
+        >
+          <ThankYou {...ThankYouProps} />
+        </motion.div>
+      );
+    default:
+      return (
+        <motion.div
+          animate={{
+            scale: [0.94, 1.05, 1],
+          }}
+          transition={{ duration: 0.8 }}
+        >
+          <TreeDonation {...TreeDonationProps} />
+        </motion.div>
+      );
   }
 }
