+++ src/features/donations/screens/TreeDonation.tsx
@@ -10,14 +10,16 @@ import { TreeDonationProps } from '../../common/types/donations';
 import SelectCurrencyModal from '../components/treeDonation/SelectCurrencyModal';
 import SelectTaxDeductionCountryModal from '../components/treeDonation/SelectTaxDeductionCountryModal';
 import styles from '../styles/TreeDonation.module.scss';
-import { PaymentRequestCustomButton } from '../components/PaymentRequestForm';
+import { PaymentRequestCustomButton } from '../components/paymentMethods/PaymentRequestCustomButton';
 import GiftForm from '../components/treeDonation/GiftForm';
 import DirectGiftForm from '../components/treeDonation/DirectGiftForm';
-import { payWithCard } from '../components/treeDonation/PaymentFunctions';
+import { payWithCard } from '../components/PaymentFunctions';
 import i18next from '../../../../i18n';
 import getFormatedCurrency from '../../../utils/countryCurrency/getFormattedCurrency';
 import { getFormattedNumber } from '../../../utils/getFormattedNumber';
 import { getMinimumAmountForCurrency } from '../../../utils/countryCurrency/getExchange';
+import { Elements } from '@stripe/react-stripe-js';
+import getStripe from '../../../utils/stripe/getStripe';
 
 const { useTranslation } = i18next;
 
@@ -134,7 +136,7 @@ function TreeDonation({
   };
 
   const [isCustomTrees, setIsCustomTrees] = React.useState(false);
-  const [isGiftValidated,setGiftValidated] = React.useState(false);
+  const [isGiftValidated, setGiftValidated] = React.useState(false);
   return ready ? (
     isPaymentProcessing ? (
       <PaymentProgress isPaymentProcessing={isPaymentProcessing} />
@@ -338,18 +340,22 @@ function TreeDonation({
             {((treeCost * treeCount) >= minAmt) ? !isPaymentOptionsLoading &&
               paymentSetup?.gateways?.stripe?.account &&
               currency ? (
-                <PaymentRequestCustomButton
-                  country={country}
-                  currency={currency}
-                  amount={formatAmountForStripe(
-                    treeCost * treeCount,
-                    currency.toLowerCase()
-                  )}
-                  onPaymentFunction={onPaymentFunction}
-                  continueNext={continueNext}
-                  window={window}
-                  paymentSetup={paymentSetup}
-                />
+                <Elements
+                  stripe={getStripe(paymentSetup)}>
+                  <PaymentRequestCustomButton
+                    country={country}
+                    currency={currency}
+                    amount={formatAmountForStripe(
+                      treeCost * treeCount,
+                      currency.toLowerCase()
+                    )}
+                    onPaymentFunction={onPaymentFunction}
+                    continueNext={continueNext}
+                    window={window}
+                    paymentSetup={paymentSetup}
+                  />
+                </Elements>
+
               ) : (
                 <div className={styles.actionButtonsContainer}>
                   <ButtonLoader />
