+++ src/features/donations/styles/PaymentDetails.module.scss
@@ -211,3 +211,46 @@
 :export {
   primaryFontFamily: $primaryFontFamily;
 }
+
+// For legacy donations
+.donationPaymentSection {
+  min-height: 100vh;
+  width: 100%;
+  display: flex;
+  margin-top: 0px;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  padding-top: 40px;
+  padding-bottom: 0px;
+  position: relative;
+}
+
+.leaderBoardBushImage, .leaderBoardGroupTreeImage {
+  display: none;
+}
+.leaderBoardBushImageMobile {
+  position: absolute;
+  bottom: 0px;
+  width: 100%;
+}
+
+@include smTabletView {
+    .leaderBoardBushImageMobile {
+      display: none;
+    }
+}
+@include lgLaptopView {
+  .leaderBoardBushImage {
+    display: block;
+    position: absolute;
+    bottom: 0px;
+    right: 0px;
+  }
+  .leaderBoardGroupTreeImage {
+    display: block;
+    position: absolute;
+    bottom: 0px;
+    left: 0px;
+  }
+}
