--- src/features/legacyDonations/styles/PaymentDetails.module.scss
@@ -1,251 +1,0 @@
-@import './../../../theme/theme';
-
-// Used as the entire section container
-.donationPaymentSection {
-  min-height: 100vh;
-  width: 100%;
-  display: flex;
-  margin-top: 0px;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  padding-top: 40px;
-  padding-bottom: 0px;
-  position: relative;
-}
-
-.leaderBoardBushImage, .leaderBoardGroupTreeImage {
-  display: none;
-}
-.leaderBoardBushImageMobile {
-  position: absolute;
-  bottom: 0px;
-  width: 100%;
-}
-
-@include smTabletView {
-    .leaderBoardBushImageMobile {
-      display: none;
-    }
-}
-@include lgLaptopView {
-  .leaderBoardBushImage {
-    display: block;
-    position: absolute;
-    bottom: 0px;
-    right: 0px;
-  }
-  .leaderBoardGroupTreeImage {
-    display: block;
-    position: absolute;
-    bottom: 0px;
-    left: 0px;
-  }
-}
-
-// Container of the Payment options
-.container {
-  background-color: $backgroundColor;
-  border-radius: 7px;
-  padding: 20px;
-  width: 320px;
-  margin: auto;
-  margin-top: 0px;
-}
-@include smTabletView {
-  .container {
-    width: 412px;
-    margin-top: 60px;
-  }
-  .donationPaymentSection {
-    margin-top: 60px;
-  }
-}
-
-
-.header {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-
-.headerTitle {
-  font-family: $primaryFontFamily;
-  font-weight: bold;
-  font-size: 27px;
-  color: $primaryFontColor;
-}
-.headerText {
-  margin-top: 10px;
-}
-.formRow {
-  display: flex;
-  flex-direction: row;
-  width: 100%;
-  margin-top: 20px;
-}
-
-/* Tree Count and Cost  */
-
-.finalTreeCount {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-  margin-top: 20px;
-}
-
-.totalCost {
-  font-family: $primaryFontFamily;
-  font-weight: bold;
-  font-size: 16px;
-  color: $primaryColor;
-}
-
-.totalCostText {
-  font-family: $primaryFontFamily;
-  font-weight: bold;
-  font-size: 16px;
-  color: $primaryFontColor;
-  margin-left: 6px;
-}
-
-.actionButtonsContainer {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-  margin-top: 20px;
-}
-
-.actionButtonsText {
-  font-family: $primaryFontFamily;
-  font-size: $fontSubTitleSize;
-  color: $primaryFontColor;
-  margin-left: 24px;
-  margin-right: 24px;
-}
-.continueButton {
-  background-color: $primaryColor;
-  border-radius: 6px;
-  padding: 8px 40px;
-  color: $light;
-  font-family: $primaryFontFamily;
-  font-size: $fontSubTitleSize;
-  font-weight: 600;
-  &:hover {
-    cursor: pointer;
-  }
-}
-.continueButtonDisabled {
-  background-color: $dividerColor;
-  border-radius: 6px;
-  padding: 8px 40px;
-  color: $light;
-  font-family: $primaryFontFamily;
-  font-size: $fontSubTitleSize;
-  font-weight: 600;
-  &:hover {
-    cursor: pointer;
-  }
-}
-
-.horizontalLine {
-  height: 1px;
-  background-color: $dividerColor;
-  margin-top: 20px;
-}
-
-/* Payment error */
-.paymentError {
-  font-family: $primaryFontFamily;
-  font-size: $fontSubTitleSize;
-  color: #e74c3c;
-  text-align: center;
-  font-weight: 600;
-  margin-top: 20px;
-}
-
-/* Payment Mode */
-
-.paymentModeContainer {
-  border-radius: 9px;
-  border: 1px solid $dividerColor;
-  padding: $fontSubTitleSize 12px;
-  margin-top: 20px;
-  width: 100%;
-}
-.paymentModeContainer:hover {
-  cursor: pointer;
-}
-
-.paymentModeHeader {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.paymentModeHeader > svg {
-  height: 20px;
-}
-
-.paymentModeFee {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-}
-
-.paymentModeTitle {
-  font-family: $primaryFontFamily;
-  font-size: $fontSubTitleSize;
-  color: $primaryFontColor;
-  font-weight: 600;
-  flex-grow: 1;
-  margin-left: 8px;
-}
-
-.paymentModeFeeAmount {
-  font-family: $primaryFontFamily;
-  font-size: $fontSubTitleSize;
-  color: $primaryFontColor;
-  margin-right: 4px;
-}
-
-/* Save Company */
-.saveCard {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-  margin-top: 20px;
-}
-
-.saveCardText {
-  font-family: $primaryFontFamily;
-  font-size: $fontSubTitleSize;
-  color: $primaryFontColor;
-}
-
-.mandateAcceptance {
-  font-family: $primaryFontFamily;
-  font-size: $fontSubTitleSize - 2;
-  color: $primaryFontColor;
-  margin-top: 20px;
-  margin-bottom: 20px;
-}
-
-.donationPaymentLoader{
-  border: 2px solid $primaryColor;
-  border-radius: 50%;
-  border-top: 2px solid transparent;
-  width: 16px!important;
-  height: 16px!important;
-  -webkit-animation: spin 1s linear infinite; /* Safari */
-  animation: spin 1s linear infinite;
-}
-
-:export {
-  primaryFontFamily: $primaryFontFamily;
-}
