+++ src/utils/stripe/getStripe.tsx
@@ -1,13 +1,11 @@
-/**
- * This is a singleton to ensure we only instantiate Stripe once.
- */
 import { Stripe, loadStripe } from '@stripe/stripe-js';
 
 let stripePromise: Promise<Stripe | null>;
-const getStripe = (key:any,account:any) => {
-  if (!stripePromise) {
-    stripePromise = loadStripe(key,{stripeAccount:account});
-  }
+
+const getStripe = (paymentSetup:any) => {
+  let key = paymentSetup?.gateways?.stripe?.authorization.stripePublishableKey ? paymentSetup?.gateways?.stripe?.authorization.stripePublishableKey : paymentSetup?.gateways?.stripe?.stripePublishableKey;
+  let account = paymentSetup?.gateways?.stripe?.authorization.accountId;
+  stripePromise = loadStripe(key,{stripeAccount:account});
   return stripePromise;
 };
